Move old logging code out of Obsolete namespace
