+++ FO-DICOM.Core/DicomEncoding.cs
@@ -2,7 +2,6 @@
 // Licensed under the Microsoft Public License (MS-PL).
 
 using FellowOakDicom.IO.Buffer;
-using FellowOakDicom.Log;
 using FellowOakDicom.Memory;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -361,6 +360,6 @@ namespace FellowOakDicom
 
         private static ILogger Logger =>
             Setup.ServiceProvider.GetRequiredService<ILoggerFactory>()
-                .CreateLogger(LogCategories.Encoding);
+                .CreateLogger(Log.LogCategories.Encoding);
     }
 }
