+++ FO-DICOM.Core/Imaging/Codec/DicomCodecParams.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) 2012-2021 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
-using FellowOakDicom.Log;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
@@ -18,7 +17,7 @@ namespace FellowOakDicom.Imaging.Codec
         /// </summary>
         protected DicomCodecParams()
         {
-            Logger = Setup.ServiceProvider.GetRequiredService<ILoggerFactory>().CreateLogger(LogCategories.Codec);
+            Logger = Setup.ServiceProvider.GetRequiredService<ILoggerFactory>().CreateLogger(FellowOakDicom.Log.LogCategories.Codec);
         }
 
         /// <summary>
