+++ FO-DICOM.Core/Log/FellowOakDicomLoggerProvider.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using System;
 
-namespace FellowOakDicom.Log.Obsolete
+namespace FellowOakDicom.Log
 {
     [Obsolete("Fellow Oak DICOM now supports Microsoft.Extensions.Logging")]
     internal class FellowOakDicomLoggerProvider: ILoggerProvider
