+++ FO-DICOM.Core/Network/DicomServer.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) 2012-2021 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
-using FellowOakDicom.Log;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using System;
@@ -132,7 +131,7 @@ namespace FellowOakDicom.Network
         /// <inheritdoc />
         public ILogger Logger
         {
-            get => _logger ??= _loggerFactory.CreateLogger(LogCategories.Network);
+            get => _logger ??= _loggerFactory.CreateLogger(Log.LogCategories.Network);
             set => _logger  = value;
         }
