+++ FO-DICOM.Core/Setup.cs
@@ -4,7 +4,7 @@
 using FellowOakDicom.Imaging;
 using FellowOakDicom.Imaging.Codec;
 using FellowOakDicom.IO;
-using FellowOakDicom.Log.Obsolete;
+using FellowOakDicom.Log;
 using FellowOakDicom.Memory;
 using FellowOakDicom.Network;
 using FellowOakDicom.Network.Client;
